requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary
python-dotenv